requirements.txt
transformers>=4.15.0
datasets>=2.0.0

torch>=1.9.0

spacy==3.5.4
scispacy>=0.5.0
owlready2>=0.37

python-docx>=0.8.11

"pydantic<2.0.0"

typing_extensions>=4.4.0

tiktoken>=0.4.0
protobuf>=3.13.0

sentencepiece>=0.1.96